Feature/refactor (#7)
* Refactor started

* Added SignalR

* Fixed WordDocument Creation

* Added GetScheduleByDateQueryHandler tests

* Added GetAllUsersQueryHandler, GetUserByIdQuery, GetScheduleByDateQuery, AddNewHolidayCommand and DeleteHolidayCommand tests

* Added AddNewRequestCommandHandler, DeleteRequestCommandHandler and AddNewScheduleCommandHandler tests

* Added DeleteScheduleCommandHandler tests and eventing

* Modified config.yml

* config.yml fixed

* config.yml fixed line 29

* fixed path in config

* Updated config.yml

* updated config.yml

* fixed dotnet test command in config.yml

* Updated config.yml

* Updated config.yml

* Added Serilog

* Added logging to every command and event handlers and configured Seq

+++ WorkSchedule.Application/CommandHandlers/Holidays/DeleteHolidayCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Serilog;
 using WorkSchedule.Api.Commands.Holidays;
 using WorkSchedule.Application.Data;
 using WorkSchedule.Application.Exceptions;
@@ -8,10 +9,11 @@ namespace WorkSchedule.Application.CommandHandlers.Holidays
     public class DeleteHolidayCommandHandler : IRequestHandler<DeleteHolidayCommand, Unit>
     {
         private readonly IUnitOfWork _uow;
-
-        public DeleteHolidayCommandHandler(IUnitOfWork uow)
+        private readonly ILogger _logger;
+        public DeleteHolidayCommandHandler(IUnitOfWork uow, ILogger logger)
         {
             _uow = uow;
+            _logger = logger;
         }
 
         public async Task<Unit> Handle(DeleteHolidayCommand request, CancellationToken cancellationToken)
@@ -23,8 +25,12 @@ namespace WorkSchedule.Application.CommandHandlers.Holidays
                 throw new BusinessException { ErrorCode = 599, ErrorMessages = validatorResult.Errors.Select(e => e.ErrorMessage).ToList() };
             }
             var holiday = await _uow.HolidayRepository.GetByID(request.Id);
-            _uow.HolidayRepository.Delete(holiday);
-            _uow.Save();
+            if (holiday is not null)
+            {
+                _uow.HolidayRepository.Delete(holiday);
+                _uow.Save();
+                _logger.Information($"Holiday with ID: {holiday.Id} has been deleted");
+            }
             return Unit.Value;
         }
     }
